fix issues for FW versions with more than 255 params

+++ ros_ws/src/crazyswarm/src/crazyswarm_server.cpp
@@ -296,7 +296,7 @@ public:
 public:
 
   template<class T, class U>
-  void updateParam(uint8_t id, const std::string& ros_param) {
+  void updateParam(uint16_t id, const std::string& ros_param) {
       U value;
       ros::param::get(ros_param, value);
       m_cf.addSetParam<T>(id, (T)value);
